back_utils/azure_back/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

Pillow
python-dotenv
torch
torchvision
torchaudio
azure-functions
azure-data-tables
azure-storage-blob
osmnx
rasterio
openai
tiktoken
